requirements.txt
torch>=2.0.0
torchmetrics>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
hydra-core
omegaconf>=2.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm
comet-ml
